Create a method to establish a connection with server
Create a method in our ContactsProgram class to obtain access to the
server. The method will establish said connection with a JDBC url, a
username, and a password (user is prompted for the password). Using this
information, a Connection object is generated by DriverManager. We can
then presumably interact with said Connection object in order to
communicate back and forth with the PostgreSQL server.